+++ R/globals.R
@@ -1,7 +1,8 @@
 utils::globalVariables(c("RMSE", "adjR", "choice", "control", "controlID",
                          "control_coefs", "estimate", "estimate_FE", "feols",
                          "fixed", "included", "index", "is_fe", "is_residual",
-                         "label", "lower", "observed", "outside", "p", "percent",
-                         "ribbon.group", "se", "se_type", "sig", "sig.level",
-                         "statistic", "term", "upper", "value", "."))
+                         "label", "lower", "n_obs", "observed", "outside", "p",
+                         "percent", "ribbon.group", "se", "se_type", "sig",
+                         "sig.level", "statistic", "term", "upper", "value",
+                         "."))